+++ model_utils/models.py
@@ -1,21 +1,18 @@
 from django.core.exceptions import ImproperlyConfigured
-from django.db import models, transaction, router
+from django.db import models, router, transaction
+from django.db.models.functions import Now
 from django.db.models.signals import post_save, pre_save
 from django.utils.translation import gettext_lazy as _
 
 from model_utils.fields import (
     AutoCreatedField,
     AutoLastModifiedField,
-    StatusField,
     MonitorField,
+    StatusField,
     UUIDField,
 )
-from model_utils.managers import (
-    QueryManager,
-    SoftDeletableManager,
-)
+from model_utils.managers import QueryManager, SoftDeletableManager
 
-from django.db.models.functions import Now
 now = Now()
 
 
@@ -34,10 +31,12 @@ class TimeStampedModel(models.Model):
         modified field is updated even if it is not given as
         a parameter to the update field argument.
         """
-        if 'update_fields' in kwargs and 'modified' not in kwargs['update_fields']:
-            kwargs['update_fields'] += ['modified']
+        update_fields = kwargs.get('update_fields', None)
+        if update_fields:
+            kwargs['update_fields'] = set(update_fields).union({'modified'})
+
         super().save(*args, **kwargs)
-        
+
     class Meta:
         abstract = True
 
@@ -66,6 +65,18 @@ class StatusModel(models.Model):
     """
     status = StatusField(_('status'))
     status_changed = MonitorField(_('status changed'), monitor='status')
+
+    def save(self, *args, **kwargs):
+        """
+        Overriding the save method in order to make sure that
+        status_changed field is updated even if it is not given as
+        a parameter to the update field argument.
+        """
+        update_fields = kwargs.get('update_fields', None)
+        if update_fields and 'status' in update_fields:
+            kwargs['update_fields'] = set(update_fields).union({'status_changed'})
+
+        super().save(*args, **kwargs)
 
     class Meta:
         abstract = True
@@ -132,7 +143,8 @@ class SoftDeletableModel(models.Model):
     class Meta:
         abstract = True
 
-    objects = SoftDeletableManager()
+    objects = SoftDeletableManager(_emit_deprecation_warnings=True)
+    available_objects = SoftDeletableManager()
     all_objects = models.Manager()
 
     def delete(self, using=None, soft=True, *args, **kwargs):