+++ model_utils/tracker.py
@@ -1,11 +1,43 @@
 from copy import deepcopy
 from functools import wraps
 
-import django
 from django.core.exceptions import FieldError
 from django.db import models
-from django.db.models.fields.files import FileDescriptor
+from django.db.models.fields.files import FieldFile, FileDescriptor
 from django.db.models.query_utils import DeferredAttribute
+
+
+class LightStateFieldFile(FieldFile):
+    """
+    FieldFile subclass with the only aim to remove the instance from the state.
+
+    The change introduced in Django 3.1 on FieldFile subclasses results in pickling the
+    whole instance for every field tracked.
+    As this is done on the initialization of objects, a simple queryset evaluation on
+    Django 3.1+ can make the app unusable, as CPU and memory usage gets easily
+    multiplied by magnitudes.
+    """
+    def __getstate__(self):
+        """
+        We don't need to deepcopy the instance, so nullify if provided.
+        """
+        state = super().__getstate__()
+        if 'instance' in state:
+            state['instance'] = None
+        return state
+
+
+def lightweight_deepcopy(value):
+    """
+    Use our lightweight class to avoid copying the instance on a FieldFile deepcopy.
+    """
+    if isinstance(value, FieldFile):
+        value = LightStateFieldFile(
+            instance=value.instance,
+            field=value.field,
+            name=value.name,
+        )
+    return deepcopy(value)
 
 
 class DescriptorMixin:
@@ -21,7 +53,7 @@ class DescriptorMixin:
             was_deferred = True
         value = super().__get__(instance, owner)
         if was_deferred:
-            self.tracker_instance.saved_data[field_name] = deepcopy(value)
+            self.tracker_instance.saved_data[field_name] = lightweight_deepcopy(value)
         return value
 
     def _get_field_name(self):
@@ -45,11 +77,11 @@ class DescriptorWrapper:
             value = self.descriptor
         if was_deferred:
             tracker_instance = getattr(instance, self.tracker_attname)
-            tracker_instance.saved_data[self.field_name] = deepcopy(value)
+            tracker_instance.saved_data[self.field_name] = lightweight_deepcopy(value)
         return value
 
     def __set__(self, instance, value):
-        initialized = hasattr(instance, '_instance_intialized')
+        initialized = hasattr(instance, '_instance_initialized')
         was_deferred = self.field_name in instance.get_deferred_fields()
 
         # Sentinel attribute to detect whether we are already trying to
@@ -87,15 +119,90 @@ class FullDescriptorWrapper(DescriptorWrapper):
         self.descriptor.__delete__(obj)
 
 
+class FieldsContext:
+    """
+    A context manager for tracking nested reset fields contexts.
+
+    If tracked fields is mentioned in more than one FieldsContext, it's state
+    is being reset on exiting last context that mentions that field.
+
+    >>> with fields_context(obj.tracker, 'f1', state=state):
+    ...     with fields_context(obj.tracker, 'f1', 'f2', state=state):
+    ...         obj.do_something_useful()
+    ...     # f2 is reset after inner context exit
+    ...     obj.do_something_else()
+    ... # f1 is reset after outer context exit
+    >>>
+
+    * Note that fields are counted by passing same state dict
+    * FieldsContext is instantiated using FieldInstanceTracker (`obj.tracker`)
+    * Different objects has own state stack
+
+    """
+
+    def __init__(self, tracker, *fields, state=None):
+        """
+        :param tracker: FieldInstanceTracker instance to be reset after
+            context exit
+        :param fields: a list of field names to be tracked in current context
+        :param state: shared state dict used to count number of field
+            occurrences in context stack.
+
+        On context enter each field mentioned in `fields` has +1 in shared
+        state, and on exit it receives -1. Fields that have zero after context
+        exit are reset in tracker instance.
+        """
+        if state is None:
+            state = {}
+        self.tracker = tracker
+        self.fields = fields
+        self.state = state
+
+    def __enter__(self):
+        """
+        Increments tracked fields occurrences count in shared state.
+        """
+        for f in self.fields:
+            self.state.setdefault(f, 0)
+            self.state[f] += 1
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """
+        Decrements tracked fields occurrences count in shared state.
+
+        If any field has no more occurrences in shared state, this field is
+        being reset by tracker.
+        """
+        reset_fields = []
+        for f in self.fields:
+            self.state[f] -= 1
+            if self.state[f] == 0:
+                reset_fields.append(f)
+                del self.state[f]
+        if reset_fields:
+            self.tracker.set_saved_fields(fields=reset_fields)
+
+
 class FieldInstanceTracker:
     def __init__(self, instance, fields, field_map):
         self.instance = instance
         self.fields = fields
         self.field_map = field_map
+        self.context = FieldsContext(self, *self.fields)
+
+    def __enter__(self):
+        return self.context.__enter__()
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        return self.context.__exit__(exc_type, exc_val, exc_tb)
+
+    def __call__(self, *fields):
+        return FieldsContext(self, *fields, state=self.context.state)
 
     @property
     def deferred_fields(self):
-        return self.instance._deferred_fields if django.VERSION < (1, 10) else self.instance.get_deferred_fields()
+        return self.instance.get_deferred_fields()
 
     def get_field_value(self, field):
         return getattr(self.instance, self.field_map[field])
@@ -110,7 +217,7 @@ class FieldInstanceTracker:
 
         # preventing mutable fields side effects
         for field, field_value in self.saved_data.items():
-            self.saved_data[field] = deepcopy(field_value)
+            self.saved_data[field] = lightweight_deepcopy(field_value)
 
     def current(self, fields=None):
         """Returns dict of current values for all tracked fields"""
@@ -151,7 +258,7 @@ class FieldInstanceTracker:
             else:
                 current_value = self.get_field_value(field)
                 self.instance.refresh_from_db(fields=[field])
-                self.saved_data[field] = deepcopy(self.get_field_value(field))
+                self.saved_data[field] = lightweight_deepcopy(self.get_field_value(field))
                 setattr(self.instance, self.field_map[field], current_value)
 
         return self.saved_data.get(field)
@@ -185,7 +292,7 @@ class FieldInstanceTracker:
                 field_tracker = FileDescriptorTracker(field_obj.field)
                 setattr(self.instance.__class__, field, field_tracker)
             else:
-                field_tracker = DeferredAttributeTracker(field, type(self.instance))
+                field_tracker = DeferredAttributeTracker(field)
                 setattr(self.instance.__class__, field, field_tracker)
 
 
@@ -195,6 +302,20 @@ class FieldTracker:
 
     def __init__(self, fields=None):
         self.fields = fields
+
+    def __call__(self, func=None, fields=None):
+        def decorator(f):
+            @wraps(f)
+            def inner(obj, *args, **kwargs):
+                tracker = getattr(obj, self.attname)
+                field_list = tracker.fields if fields is None else fields
+                with tracker(*field_list):
+                    return f(obj, *args, **kwargs)
+
+            return inner
+        if func is None:
+            return decorator
+        return decorator(func)
 
     def get_field_map(self, cls):
         """Returns dict mapping fields names to model attribute names"""
@@ -213,12 +334,11 @@ class FieldTracker:
         if self.fields is None:
             self.fields = (field.attname for field in sender._meta.fields)
         self.fields = set(self.fields)
-        if django.VERSION >= (1, 10):
-            for field_name in self.fields:
-                descriptor = getattr(sender, field_name)
-                wrapper_cls = DescriptorWrapper.cls_for_descriptor(descriptor)
-                wrapped_descriptor = wrapper_cls(field_name, descriptor, self.attname)
-                setattr(sender, field_name, wrapped_descriptor)
+        for field_name in self.fields:
+            descriptor = getattr(sender, field_name)
+            wrapper_cls = DescriptorWrapper.cls_for_descriptor(descriptor)
+            wrapped_descriptor = wrapper_cls(field_name, descriptor, self.attname)
+            setattr(sender, field_name, wrapped_descriptor)
         self.field_map = self.get_field_map(sender)
         models.signals.post_init.connect(self.initialize_tracker)
         self.model_class = sender
@@ -231,7 +351,7 @@ class FieldTracker:
         tracker = self.tracker_class(instance, self.fields, self.field_map)
         setattr(instance, self.attname, tracker)
         tracker.set_saved_fields()
-        instance._instance_intialized = True
+        instance._instance_initialized = True
 
     def patch_save(self, model):
         self._patch(model, 'save_base', 'update_fields')
@@ -242,21 +362,17 @@ class FieldTracker:
 
         @wraps(original)
         def inner(instance, *args, **kwargs):
-            ret = original(instance, *args, **kwargs)
             update_fields = kwargs.get(fields_kwarg)
-            if not update_fields and update_fields is not None:  # () or []
-                fields = update_fields
-            elif update_fields is None:
-                fields = None
+            if update_fields is None:
+                fields = self.fields
             else:
                 fields = (
                     field for field in update_fields if
                     field in self.fields
                 )
-            getattr(instance, self.attname).set_saved_fields(
-                fields=fields
-            )
-            return ret
+            tracker = getattr(instance, self.attname)
+            with tracker(*fields):
+                return original(instance, *args, **kwargs)
 
         setattr(model, method, inner)