+++ tests/models.py
@@ -1,32 +1,23 @@
 import django
 from django.db import models
-from django.db.models.query_utils import DeferredAttribute
 from django.db.models import Manager
-from django.utils.translation import ugettext_lazy as _
+from django.db.models.query_utils import DeferredAttribute
+from django.utils.translation import gettext_lazy as _
 
 from model_utils import Choices
-from model_utils.fields import (
-    SplitField,
-    MonitorField,
-    StatusField,
-    UUIDField,
-)
-from model_utils.managers import (
-    QueryManager,
-    InheritanceManager,
-    JoinManagerMixin
-)
+from model_utils.fields import MonitorField, SplitField, StatusField, UUIDField
+from model_utils.managers import InheritanceManager, JoinManagerMixin, QueryManager
 from model_utils.models import (
+    SaveSignalHandlingModel,
     SoftDeletableModel,
     StatusModel,
     TimeFramedModel,
     TimeStampedModel,
     UUIDModel,
-    SaveSignalHandlingModel,
 )
+from model_utils.tracker import FieldTracker, ModelTracker
 from tests.fields import MutableField
 from tests.managers import CustomSoftDeleteManager
-from model_utils.tracker import FieldTracker, ModelTracker
 
 
 class InheritanceManagerTestRelated(models.Model):
@@ -89,7 +80,7 @@ class InheritanceManagerTestChild4(InheritanceManagerTestParent):
 
 
 class TimeStamp(TimeStampedModel):
-    pass
+    test_field = models.PositiveSmallIntegerField(default=0)
 
 
 class TimeFrame(TimeFramedModel):
@@ -456,3 +447,13 @@ class CustomNotPrimaryUUIDModel(models.Model):
 
 class SaveSignalHandlingTestModel(SaveSignalHandlingModel):
     name = models.CharField(max_length=20)
+
+
+class TimeStampWithStatusModel(TimeStampedModel, StatusModel):
+    STATUS = Choices(
+        ("active", _("active")),
+        ("deleted", _("deleted")),
+        ("on_hold", _("on hold")),
+    )
+
+    test_field = models.PositiveSmallIntegerField(default=0)