+++ tests/test_models/test_timestamped_model.py
@@ -1,10 +1,9 @@
 from datetime import datetime, timedelta
 
+from django.test import TestCase
 from freezegun import freeze_time
 
-from django.test import TestCase
-
-from tests.models import TimeStamp
+from tests.models import TimeStamp, TimeStampWithStatusModel
 
 
 class TimeStampedModelTests(TestCase):
@@ -90,31 +89,62 @@ class TimeStampedModelTests(TestCase):
         self.assertEqual(t1.created, different_date2)
         self.assertNotEqual(t1.modified, different_date2)
         self.assertNotEqual(t1.modified, different_date)
-    
-    def test_save_with_update_fields_overrides_modified_provided(self):
-        '''
+
+    def test_save_with_update_fields_overrides_modified_provided_within_a(self):
+        """
         Tests if the save method updated modified field
         accordingly when update_fields is used as an argument
         and modified is provided
-        '''
-        with freeze_time(datetime(2020,1,1)):
+        """
+        tests = (
+            ['modified'],  # list
+            ('modified',),  # tuple
+            {'modified'},  # set
+        )
+
+        for update_fields in tests:
+            with self.subTest(update_fields=update_fields):
+                with freeze_time(datetime(2020, 1, 1)):
+                    t1 = TimeStamp.objects.create()
+
+                with freeze_time(datetime(2020, 1, 2)):
+                    t1.save(update_fields=update_fields)
+                self.assertEqual(t1.modified, datetime(2020, 1, 2))
+
+    def test_save_is_skipped_for_empty_update_fields_iterable(self):
+        tests = (
+            [],  # list
+            (),  # tuple
+            set(),  # set
+        )
+
+        for update_fields in tests:
+            with self.subTest(update_fields=update_fields):
+                with freeze_time(datetime(2020, 1, 1)):
+                    t1 = TimeStamp.objects.create()
+
+                with freeze_time(datetime(2020, 1, 2)):
+                    t1.test_field = 1
+                    t1.save(update_fields=update_fields)
+
+                t1.refresh_from_db()
+                self.assertEqual(t1.test_field, 0)
+                self.assertEqual(t1.modified, datetime(2020, 1, 1))
+
+    def test_save_updates_modified_value_when_update_fields_explicitly_set_to_none(self):
+        with freeze_time(datetime(2020, 1, 1)):
             t1 = TimeStamp.objects.create()
-        
-        with freeze_time(datetime(2020,1,2)):
-            t1.save(update_fields=['modified'])
-        
-        self.assertEqual(t1.modified, datetime(2020,1,2))
-        
-    def test_save_with_update_fields_overrides_modified_not_provided(self):
-        '''
-        Tests if the save method updated modified field
-        accordingly when update_fields is used as an argument
-        and modified is not provided
-        '''
-        with freeze_time(datetime(2020,1,1)):
-            t1 = TimeStamp.objects.create()
-            
-        with freeze_time(datetime(2020,1,2)):
-            t1.save(update_fields=[])
-            
-        self.assertEqual(t1.modified, datetime(2020,1,2))
+
+        with freeze_time(datetime(2020, 1, 2)):
+            t1.save(update_fields=None)
+
+        self.assertEqual(t1.modified, datetime(2020, 1, 2))
+
+    def test_model_inherit_timestampmodel_and_statusmodel(self):
+        with freeze_time(datetime(2020, 1, 1)):
+            t1 = TimeStampWithStatusModel.objects.create()
+
+        with freeze_time(datetime(2020, 1, 2)):
+            t1.save(update_fields=['test_field', 'status'])
+
+        self.assertEqual(t1.modified, datetime(2020, 1, 2))